changing task delete from get request to post request

+++ components/mediation-ui/org.wso2.carbon.task.ui/src/main/resources/web/task/deletetask.jsp
@@ -1,20 +1,20 @@
 <!--
- ~ Copyright (c) 2005-2010, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- ~
- ~ WSO2 Inc. licenses this file to you under the Apache License,
- ~ Version 2.0 (the "License"); you may not use this file except
- ~ in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~    http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing,
- ~ software distributed under the License is distributed on an
- ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- ~ KIND, either express or implied.  See the License for the
- ~ specific language governing permissions and limitations
- ~ under the License.
- -->
+~ Copyright (c) 2005-2010, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+~
+~ WSO2 Inc. licenses this file to you under the Apache License,
+~ Version 2.0 (the "License"); you may not use this file except
+~ in compliance with the License.
+~ You may obtain a copy of the License at
+~
+~ http://www.apache.org/licenses/LICENSE-2.0
+~
+~ Unless required by applicable law or agreed to in writing,
+~ software distributed under the License is distributed on an
+~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+~ KIND, either express or implied. See the License for the
+~ specific language governing permissions and limitations
+~ under the License.
+-->
 <%@ page import="org.wso2.carbon.task.ui.internal.TaskClientConstants" %>
 <%@ page import="org.wso2.carbon.task.ui.internal.TaskManagementClient" %>
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
@@ -23,28 +23,33 @@
 <fmt:bundle basename="org.wso2.carbon.task.ui.i18n.Resources">
 
     <%
-        TaskManagementClient client;
-        try {
+        //ignore methods other than post
+        if (!request.getMethod().equalsIgnoreCase("POST")) {
+            response.sendError(405);
+            return;
+        } else {
+            TaskManagementClient client;
+            try {
 
-            String name = request.getParameter("taskName");
-            if (name == null || "".equals(name)) {
-                throw new ServletException("Task name is empty");
-            }
+                String name = request.getParameter("taskName");
+                if (name == null || "".equals(name)) {
+                    throw new ServletException("Task name is empty");
+                }
 
-            String group = request.getParameter("taskGroup");
-            if (group == null || "".equals(group)) {
-                throw new ServletException("Task group is empty");
-            }
+                String group = request.getParameter("taskGroup");
+                if (group == null || "".equals(group)) {
+                    throw new ServletException("Task group is empty");
+                }
 
-            client = TaskManagementClient.getInstance(config, session);
-            name = name.trim();
-            client.deleteTaskDescription(name, group);
+                client = TaskManagementClient.getInstance(config, session);
+                name = name.trim();
+                client.deleteTaskDescription(name, group);
 //            request.getSession().removeAttribute(TaskClientConstants.TASK_KEY + name);
-        } catch (Throwable e) {
-            request.getSession().setAttribute(TaskClientConstants.EXCEPTION, e);
+            } catch (Throwable e) {
+                request.getSession().setAttribute(TaskClientConstants.EXCEPTION, e);
     %>
     <script type="text/javascript">
-        jQuery(document).ready(function() {
+        jQuery(document).ready(function () {
             CARBON.showErrorDialog('<%=e.getMessage()%>');
         });
     </script>
@@ -55,4 +60,7 @@
     <script type="text/javascript">
         forward("index.jsp");
     </script>
+    <%
+        }
+    %>
 </fmt:bundle>
